requirements.txt
langchain==0.2.14
langchain-community==0.2.12
langchain-google-genai==1.0.8
chromadb==0.5.5
pypdf==4.3.1
python-dotenv==1.0.1
tiktoken==0.7.0
pydantic>=2.0
fpdf2==2.7.9
streamlit==1.32.0
langchain-text-splitters==0.2.4
